rename `dialogs.errormessagebox` -> `dialogs.error_message`
